+++ app/api/auth/[...nextauth]/route.ts
@@ -1,70 +1,5 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import bcrypt from "bcryptjs";
-import client from "@/app/libs/prismadb"; // Correct import for prisma client
+import NextAuth from "next-auth";
+import { options } from "./options";
 
-export const authOptions: AuthOptions = {
-  secret: process.env.NEXTAUTH_SECRET, // Ensure this is in your .env file
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        const user = await client.userAccount.findFirst({
-          where: {
-            OR: [
-              { email: credentials?.email },
-              { username: credentials?.email },
-            ],
-          },
-        });
-
-        if (user && credentials?.password) {
-          const isValidPassword = await bcrypt.compare(
-            credentials.password,
-            user.password
-          );
-          if (isValidPassword) {
-            return user;
-          }
-        }
-        return null; // Return null if authentication fails
-      },
-    }),
-  ],
-  adapter: PrismaAdapter(client),
-  session: {
-    strategy: "jwt", // Use JWT for session management
-  },
-  pages: {
-    signIn: "",
-  },
-  callbacks: {
-    async session({ session, token }) {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.sub, // Attach user ID from token
-          image: token.picture || "", // Attach user image from token
-          role: token.user || "",
-        },
-      };
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id; // Attach the user ID to the JWT token
-        token.role = user.role;
-      }
-      return token;
-    },
-  },
-};
-
-// Export the NextAuth handler for both GET and POST requests
-const handler = NextAuth(authOptions);
+const handler = NextAuth(options);
 export { handler as GET, handler as POST };
